+++ Tests/Unit/Controller/EventRegistrationControllerTest.php
@@ -7,8 +7,11 @@ namespace OliverKlee\Seminars\Tests\Unit\Controller;
 use Nimut\TestingFramework\MockObject\AccessibleMockObjectInterface;
 use Nimut\TestingFramework\TestCase\UnitTestCase;
 use OliverKlee\Seminars\Controller\EventRegistrationController;
+use OliverKlee\Seminars\Domain\Model\Event\SingleEvent;
+use OliverKlee\Seminars\Service\RegistrationGuard;
 use PHPUnit\Framework\MockObject\MockObject;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
+use TYPO3\CMS\Extbase\Mvc\Exception\StopActionException;
 use TYPO3\CMS\Fluid\View\TemplateView;
 
 /**
@@ -20,6 +23,11 @@ final class EventRegistrationControllerTest extends UnitTestCase
      * @var EventRegistrationController&MockObject&AccessibleMockObjectInterface
      */
     private $subject;
+
+    /**
+     * @var RegistrationGuard&MockObject
+     */
+    private $registrationGuardMock;
 
     /**
      * @var TemplateView&MockObject
@@ -36,6 +44,9 @@ final class EventRegistrationControllerTest extends UnitTestCase
             ['redirect', 'forward', 'redirectToUri']
         );
         $this->subject = $subject;
+
+        $this->registrationGuardMock = $this->createMock(RegistrationGuard::class);
+        $this->subject->injectRegistrationGuard($this->registrationGuardMock);
 
         $this->viewMock = $this->createMock(TemplateView::class);
         $this->subject->_set('view', $this->viewMock);
@@ -57,7 +68,9 @@ final class EventRegistrationControllerTest extends UnitTestCase
         $pageUid = 42;
         $this->subject->_set('settings', ['pageForMissingEvent' => (string)$pageUid]);
 
-        $this->subject->expects(self::once())->method('redirect')->with(null, null, null, [], $pageUid);
+        $this->subject->expects(self::once())->method('redirect')->with(null, null, null, [], $pageUid)
+            ->willThrowException(new StopActionException('redirectToUri', 1476045828));
+        $this->expectException(StopActionException::class);
 
         $this->subject->checkPrerequisitesAction();
     }
@@ -70,7 +83,9 @@ final class EventRegistrationControllerTest extends UnitTestCase
         $pageUid = 42;
         $this->subject->_set('settings', ['pageForMissingEvent' => (string)$pageUid]);
 
-        $this->subject->expects(self::once())->method('redirect')->with(null, null, null, [], $pageUid);
+        $this->subject->expects(self::once())->method('redirect')->with(null, null, null, [], $pageUid)
+            ->willThrowException(new StopActionException('redirectToUri', 1476045828));
+        $this->expectException(StopActionException::class);
 
         $this->subject->checkPrerequisitesAction(null);
     }
@@ -78,9 +93,26 @@ final class EventRegistrationControllerTest extends UnitTestCase
     /**
      * @test
      */
+    public function checkPrerequisitesActionForNoRegistrationPossibleAtAllForwardsToDenyRegistrationAction(): void
+    {
+        $event = new SingleEvent();
+        $this->registrationGuardMock->expects(self::once())->method('isRegistrationPossibleAtAnyTimeAtAll')
+            ->with($event)->willReturn(false);
+
+        $this->subject->expects(self::once())->method('forward')
+            ->with('denyRegistration', null, null, ['warningMessageKey' => 'noRegistrationPossibleAtAll'])
+            ->willThrowException(new StopActionException('forward', 1476045801));
+        $this->expectException(StopActionException::class);
+
+        $this->subject->checkPrerequisitesAction($event);
+    }
+
+    /**
+     * @test
+     */
     public function denyRegistrationActionPassesProvidedWarningMessageKeyToView(): void
     {
-        $warningMessageKey = 'not-possible';
+        $warningMessageKey = 'noRegistrationPossibleAtAll';
         $this->viewMock->expects(self::once())->method('assign')->with('warningMessageKey', $warningMessageKey);
 
         $this->subject->denyRegistrationAction($warningMessageKey);